refactor: Remove PaymentSuccessPage component and related logic
- Deleted PaymentSuccessPage component and its associated logic from the payment success handling.
- Cleaned up unused imports and functionality to streamline the codebase.
